[Core] Move command line parser implementation
